Move defaultBackupManager to BackupWatcher
